Fix link error when libpng/libjpeg isn't installed.
If libPNG and/or JPEG libraries are not installed and Perl was compiled
with -D_FORTIFY_SOURCE, the linker will fail with a mismatched errno
error. This is caused by GCC outputting a warning when Perl/Tk's build system
picks up on the _FORTIFY_SOURCE option but not the optimization option.
The configure script checks the compiler output against the empty
string, so the warning causes it to assume there is no errno.h on the
system.

Fixes #109.